backend/requirements.txt
Flask==2.0.1
Flask-MySQLdb==1.0.1
Flask-Cors==3.0.10
mysqlclient==2.0.1
requests==2.26.0
Flask-Session==0.4.0